Backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
lancedb==0.5.3
langchain==0.1.12
langchain-google-genai==0.0.11
google-generativeai==0.3.2
sentence-transformers==2.5.1
pandas==2.2.1
numpy==1.26.4
xlsxwriter==3.2.0
apscheduler==3.10.4 